SoftProj/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
PyJWT
django-filter
